chore: renamin contact with upsert references
