+++ core/api/src/app/contacts/index.ts
@@ -1,2 +1,2 @@
-export * from "./upsert-contact"
-export * from "./upsert-intraledger-contact"
+export * from "./contact-create"
+export * from "./intraledger-contact-create"
