+++ core/api/src/graphql/public/mutations.ts
@@ -41,7 +41,7 @@ import UserLoginMutation from "@/graphql/public/root/mutation/user-login"
 import AccountDeleteMutation from "@/graphql/public/root/mutation/account-delete"
 import AccountUpdateDefaultWalletIdMutation from "@/graphql/public/root/mutation/account-update-default-wallet-id"
 import AccountUpdateDisplayCurrencyMutation from "@/graphql/public/root/mutation/account-update-display-currency"
-import ContactCreateMutation from "@/graphql/public/root/mutation/account-contact-upsert"
+import ContactCreateMutation from "@/graphql/public/root/mutation/contact-create"
 import FeedbackSubmitMutation from "@/graphql/public/root/mutation/feedback-submit"
 import OnChainPaymentSendMutation from "@/graphql/public/root/mutation/onchain-payment-send"
 import OnChainPaymentSendAllMutation from "@/graphql/public/root/mutation/onchain-payment-send-all"
